chore: move EnrichedText into native directory
