+++ src/index.native.tsx
@@ -1,5 +1,5 @@
+// EnrichedTextInput
 export { EnrichedTextInput } from './native/EnrichedTextInput';
-export * from './EnrichedText';
 export type {
   EnrichedInputStyle,
   EnrichedTextInputProps,
@@ -19,6 +19,13 @@ export type {
   EnrichedTextInputInstance,
   ContextMenuItem,
   OnChangeMentionEvent,
+} from './types';
+
+// EnrichedText
+export { EnrichedText } from './native/EnrichedText';
+export type {
+  EnrichedTextProps,
+  EnrichedTextInstance,
   EnrichedTextHtmlStyle,
   OnMentionPressEvent,
   OnLinkPressEvent,
